feat: Add on-demand progress aggregation for multi-client downloads
This commit introduces a new feature that allows for on-demand progress aggregation for multi-client downloads. This is achieved by moving the aggregate_progress function from the MultiClientDownloader struct to a standalone function. This function can now be called when progress is requested, allowing for more accurate and up-to-date progress reports.

The commit also includes changes to the load_progress function in the Job struct to support this new feature. Now, when load_progress is called for a multi-client download job, it will first attempt to aggregate progress from all client subdirectories before reading the progress file.

Finally, the commit updates the build_job_info function in jobs_manager.rs to pass the node_env to the load_progress function, enabling it to perform the on-demand progress aggregation.

This feature improves the user experience by providing more accurate progress reports for multi-client downloads.

The version of the package is also updated from 2.18.9 to 2.18.10.